# OpenAPI.yaml
openapi: 3.0.1
info:
  title: API Vulpes
  description: ""
  contact:
    name: Kaique de Miranda
    email: kbdemiranda@hotmail.com
servers:
- url: http://127.0.0.1:8080/api
  description: Generated server url
...
paths:
  /usuarios/{id}:
    get:
      tags:
      - Usuários
      summary: Buscar usuário por ID
      operationId: buscarUsuarioPorId
      parameters:
      - name: id
        in: path
        description: ID do usuário
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "404":
          description: Usuário não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Usuário encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Usuario'
      security:
      - bearer-key: []
    put:
      tags:
      - Usuários
      summary: Atualizar usuário
      operationId: atualizarUsuario
      parameters:
      - name: id
        in: path
        description: ID do usuário
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        "404":
          description: Usuário não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "201":
          description: Usuário atualizado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Usuario'
      security:
      - bearer-key: []
    delete:
      tags:
      - Usuários
      summary: Excluir usuário
      operationId: excluirUsuario
      parameters:
      - name: id
        in: path
        description: ID do usuário
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "404":
          description: Usuário não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "204":
          description: Usuário excluído
      security:
      - bearer-key: []
  /usuarios/{id}/senha:
    put:
      tags:
      - Usuários
      summary: Atualizar senha do usuário
      operationId: atualizarSenha
      parameters:
      - name: id
        in: path
        description: ID do usuário
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        "404":
          description: Usuário não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "204":
          description: Senha atualizada
      security:
      - bearer-key: []
  /usuarios/{id}/perfis:
    put:
      tags:
      - Usuários
      summary: Atualizar perfis do usuário
      operationId: atualizarPerfis
      parameters:
      - name: id
        in: path
        description: ID do usuário
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        "404":
          description: Usuário não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "204":
          description: Perfis atualizados
      security:
      - bearer-key: []
...
  /usuarios:
    get:
      tags:
      - Usuários
      summary: Buscar usuário por email
      operationId: buscarUsuarioPorEmail
      parameters:
      - name: email
        in: query
        description: Email do usuário
        required: true
        schema:
          type: string
        example: user@example.com
      responses:
        "404":
          description: Usuário não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Usuário encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Usuario'
      security:
      - bearer-key: []
    post:
      tags:
      - Usuários
      summary: Cadastrar usuário
      description: Cria um novo usuário
      operationId: cadastrarUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        "201":
          description: Usuário criado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Usuario'
        "400":
          description: Dados inválidos
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer-key: []
...
components:
  schemas:
    Usuario:
      required:
      - email
      - nome
      - perfis_id
      - senha
      - sobrenome
      type: object
      properties:
        id:
          type: integer
          description: Identificador do usuário
          format: int64
          example: 42
        nome:
          type: string
          description: Nome
          example: Kaique
        sobrenome:
          type: string
          description: Sobrenome
          example: Miranda
        email:
          type: string
          description: Email
          example: kaique@example.com
        senha:
          type: string
          description: Senha do usuário
          writeOnly: true
          example: S3nh@Segura
        perfis_id:
          type: array
          description: IDs dos perfis do usuário
          example:
          - 1
          - 2
          items:
            type: integer
            description: IDs dos perfis do usuário
            format: int64
      description: Dados de um usuário do sistema
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: Data/hora do erro (UTC)
          format: date-time
        status:
          type: integer
          description: Código HTTP do erro
          format: int32
          example: 404
        error:
          type: string
          description: Tipo de erro
          example: VulpesException
        message:
          type: string
          description: Mensagem explicativa do erro
          example: Plataforma não encontrada
        path:
          type: string
          description: Caminho da requisição
          example: /plataformas/999
      description: Estrutura padronizada de erros da API
...
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT